+++ graspit/handle_shipment.py
@@ -5,9 +5,8 @@ from graspit.services.SchedularService.center import start_service
 from graspit.services.DBService.shared import db_path
 from graspit.services.DBService.lifecycle import init_tortoise_orm, close_connection
 from graspit.services.SchedularService.lifecycle import start_loop
-from graspit.services.DBService.health import check_version
+from graspit.services.DBService.health import check_version, initiate_migration
 from graspit.services.DBService.models.config_base import ExportBase, RunBase
-from graspit.services.DBService import DB_VERSION
 from tortoise.functions import Max
 from os.path import relpath
 from click import secho
@@ -119,15 +118,15 @@ def db_version():
 
 
 @db_version.command()
-def check():
-    expected, actual, has_failed = run_async(check_version())
+@argument("path", nargs=1, type=C_Path(exists=True, dir_okay=True), required=True)
+def check(path: str):
+    return check_version(db_path(Path(path)))
 
-    matched = expected == actual
-    secho(
-        f"Expected: {expected} but found: {actual}",
-        fg="red" if not matched else "green",
-    )
-    secho(f"Please run graspit ")
+
+@db_version.command()
+@argument("path", nargs=1, type=C_Path(exists=True, dir_okay=True), required=True)
+def migrate(path: str):
+    return initiate_migration(db_path(Path(path)))
 
 
 if __name__ == "__main__":