+++ graspit-reporters/packages/wdio-graspit-reporter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # wdio-graspit-reporter
 
+## 1.4.0
+
+### Minor Changes
+
+- renamed browser to entity in the sessionbase, made changes to replicate the same on UI side and we have moved sending the details reg. session from register to mark in webdriverio
+
+### Patch Changes
+
+- Updated dependencies
+  - graspit@2.2.0
+  - graspit-commons@1.4.0
+
 ## 1.3.0
 
 ### Minor Changes